codegen: unify IntegerSpec::NamedValue / EnumeratedSpec::Value (#488)
* codegen: unify IntegerSpec::NamedValue / EnumeratedSpec::Value

Both were {asn1_name, value} pairs from the same AST source
(ast::EnumValue::number, int64_t), but EnumeratedSpec::Value used
long while IntegerSpec::NamedValue used int64_t. On LLP64 platforms
(long is 32-bit -- MSVC, all Windows targets) an ENUMERATED value
above ~2^31 silently truncated via Generator.cpp's static_cast<long>.

Unified to one namespace-scope NamedValue{asn1_name, int64_t value}
struct shared by both specs (X.680 SS19.5/SS20.6 named numbers are the
same concept). Pure consolidation elsewhere -- CppBackend/RustBackend
call sites use std::format, which is integer-width-agnostic.

Closes #485.

* codegen: clarify NamedValue::value isn't a BER tag

Addresses review comment on PR #488.

+++ compiler/src/codegen/Generator.cpp
@@ -587,11 +587,11 @@ static EnumeratedSpec build_enumerated_spec(const ast::TypeDef& def,
     spec.extensible = false;
     spec.root_count  = 0;
 
-    long auto_val = 0;
+    int64_t auto_val = 0;
     bool past_ext = false;
     for (const auto& ev : def.enum_values) {
         if (ev.name == "...") { spec.extensible = true; past_ext = true; continue; }
-        long v = static_cast<long>(ev.number.value_or(auto_val));
+        int64_t v = ev.number.value_or(auto_val);
         spec.values.push_back({ev.name, v});
         auto_val = v + 1;
         if (!past_ext) ++spec.root_count;
